+++ app/src/main/java/dev/alllexey/itmowidgets/feature/settings/domain/CustomSpoilerRepository.kt
@@ -1,7 +1,0 @@
-package dev.alllexey.itmowidgets.feature.settings.domain
-
-interface CustomSpoilerRepository {
-    suspend fun hasImage(): Boolean
-    suspend fun saveImage(sourceUri: String): Boolean
-    suspend fun resetImage(): Boolean
-}